pyproject.toml
[project]
name = "shinyapps-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datetime>=5.5",
    "ipykernel>=6.29.5",
    "japanize-matplotlib>=1.1.3",
    "jinja2>=3.1.5",
    "mplfinance>=0.12.10b0",
    "pandas>=2.2.3",
    "pathlib>=1.0.1",
    "plotly>=5.24.1",
    "requests>=2.32.3",
    "scipy>=1.14.1",
    "seaborn>=0.13.2",
    "shiny>=1.1.0",
    "shinywidgets>=0.3.4",
    "yfinance>=0.2.48",
]
